serialize + ctrl click

+++ ne_node.js
@@ -3,10 +3,29 @@ class NENode {
         this.scene = scene
         this.title = title
 
+        // get first free id
+        var curr_id = 0
+        var ids = new Array()
+        for (var i = 0; i < this.scene.nodes.length; i++){
+            var node = this.scene.nodes[i]
+            ids.push(node.id)
+        }
+        ids.every(function (a) {
+            if (curr_id === a) {
+                curr_id = a + 1;
+                return true;
+            }
+        });
+        this.id = curr_id
+
         this.ports = new Array()
 
         this.graphics_node = new NEGraphicsNode(scene, 20, 20)
         this.graphics_node.node = this
+
+        this.can_be_deleted = true
+        this.can_be_moved = true
+        this.can_be_selected = true
 
         var test_input = new NEPort(this.scene, this, 1, false)
         test_input.graphics_port.x_offset = this.graphics_node.x
@@ -24,6 +43,20 @@ class NENode {
         this.ports.push(test_output)
     }
 
+    select(scene) {
+        if(this.can_be_selected && !this.graphics_node.selected){
+
+            this.graphics_node.select()
+            scene.selected_nodes.push(this)
+        }
+    }
+
+    unselect(scene) {
+        if(this.graphics_node.selected){
+            this.graphics_node.selected = false
+        }
+    }
+
     draw() {
         this.graphics_node.draw()
 
@@ -34,10 +67,36 @@ class NENode {
 
     }
 
-    delete() {
-        for (var i = 0; i < this.ports.length; i++){
-            this.ports[i].delete()
+    delete(scene) {
+        if(this.can_be_deleted){
+            for (var i = 0; i < this.ports.length; i++){
+                this.ports[i].delete()
+            }
+
+            var idx = scene.nodes.indexOf(this)
+            scene.nodes.splice(idx, 1)
+        }else{
+            this.unselect()
         }
+    }
+
+    serialize() {
+        var ports_arr = new Array()
+        for (var i = 0; i < this.ports.length; i++){
+            var port_str = this.ports[i].serialize()
+            ports_arr.push(port_str)
+        }
+        var serialize_str = {
+            id: this.id,
+            title: this.title,
+            ports: ports_arr,
+            can_be_deleted: this.can_be_deleted,
+            can_be_selected: this.can_be_selected,
+            can_be_moved: this.can_be_moved,
+            x: this.graphics_node.x,
+            y: this.graphics_node.y
+        }
+        return serialize_str
     }
 }
 
@@ -61,7 +120,6 @@ class NEGraphicsNode {
     }
 
     select() {
-
         this.selected = true
         this.initialselectionpos.x = this.x
         this.initialselectionpos.y = this.y
